Move all curious tests into the same place
